requirements.txt
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
pandas==2.3.2
jinja2==3.1.6
apscheduler==3.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
